chore(project): schema validation for config module, tests moved
installed Joi
configmodule now validates the env variables
moved auth specs to a separate folder
