+++ backend/package.json
@@ -39,6 +39,7 @@
     "class-validator": "^0.14.0",
     "dotenv": "^16.1.4",
     "dotenv-cli": "^7.2.1",
+    "joi": "^17.10.0",
     "passport": "^0.6.0",
     "passport-jwt": "^4.0.1",
     "passport-local": "^1.0.0",
