+++ backend/src/auth/test/auth.controller.spec.ts
@@ -1,5 +1,5 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { AuthController } from './auth.controller';
+import { AuthController } from '../auth.controller';
 
 describe('AuthController', () => {
   let controller: AuthController;
